Changes for Task 1

+++ Task 1/Task 1/Product.cs
@@ -1,7 +1,6 @@
 ﻿class Product
 {
     public string name;
-    public int price, weight;
+    public decimal price, weight;
     public Product(string n, int p, int w) { name = n; price = p; weight = w; }
-
 }
